+++ include/doiplib/eid_vehicle_id_request.h
@@ -28,7 +28,7 @@ namespace DoipLib
         /// @param protocolVersion DoIP ISO protocol version
         /// @param eid Entity ID filter
         EidVehicleIdRequest(
-            uint16_t protocolVersion, std::array<uint8_t, cEidSize> eid);
+            uint8_t protocolVersion, std::array<uint8_t, cEidSize> eid);
 
         /// @brief Get EID filter
         /// @return Entity ID
