+++ samples/vehicle_id_message_handler.cpp
@@ -7,7 +7,7 @@ namespace DoipLib
     const uint8_t VehicleIdMessageHandler::cFurtherAction;
 
     VehicleIdMessageHandler::VehicleIdMessageHandler(
-        uint16_t protocolVersion,
+        uint8_t protocolVersion,
         std::string &&vin,
         uint16_t logicalAddress,
         std::array<uint8_t, cIdSize> eid,
