+++ src/vehicle_id_request.cpp
@@ -10,7 +10,7 @@ namespace DoipLib
 
 
     VehicleIdRequest::VehicleIdRequest(
-        uint16_t protocolVersion) noexcept : Message(protocolVersion, cPayloadType)
+        uint8_t protocolVersion) noexcept : Message(protocolVersion, cPayloadType)
     {
     }
